build: test working directories & build-and-test scripts

+++ runtest.cmake
@@ -1,5 +1,7 @@
 file(GLOB_RECURSE FILES ${TEST_WORKING_DIR}/**)
-file(REMOVE ${FILES})
+if (NOT "${FILES}" STREQUAL "")
+    file(REMOVE ${FILES})
+endif ()
 execute_process(COMMAND ${TEST_EXECUTABLE} RESULT_VARIABLE status)
 if(status)
     MESSAGE(FATAL_ERROR "Test executing status: ${status}")
